Created generalised binner class

+++ ClusterFinder/src/Bin.java
@@ -1,10 +1,10 @@
 public class Bin{
 
   public String name;
-  int min;
-  int max;
+  double min;
+  double max;
 
-  public Bin(String name, int min, int max){
+  public Bin(String name, double min, double max){
     this.name = name;
     this.min = min;
     this.max = max;
